Updates after Proof 2.
Page 2 - the "Our Brand" title was accidentally indented
Page 3 - please add some space between the paragraphs
Page 7 - the orange and blue colors still have the same numeric values.
Also, instead of having the colors arrange in that pattern, could you
just align them into a single column, with the color on the left and the
numeric values to the right?
Page 8 - the first sentence still has two periods at the end of the
first sentence.  Also, keep the company name in the first sentence, but
take them out in the next two sentences as they are already mentioned.
At the bottom of the page, at the very last sentence, "available" is
misspelled.
Page 10 through 12 - the measurements are still in millimeters ...
please include both measurements.
Page 10 - the rectangle holding the person's name should extend to the
far left side, and be gradient from white to orange (left-to-right).
Page 12 -there is a misspelled word: "should"